Update version and improve security features
The version of MariaDB-SQLBuilder was updated from 1.0.1 to 1.1. Keywords in the setup file were also modified, with 'orm' removed. A security fix was applied to prevent SQL injections by ensuring variables are given directly to the cursor instead of being written into SQL. This adjustment modifies the function “get_sql()” and introduces “values_for_execute” to manage variables.